+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 * The auto_translate management command has now a parameter for specifying translation mode.
 * Added support for :ref:`txt`.
+* Added trends and metrics for all objects.
 
 Weblate 4.5.3
 -------------